impl exploration stack
